// src/common/colors/colors.types.ts
export interface ColorHashI {
  black: string
  blackRussian: string
  dodgerBlue: string
  dodgerBlue25: string
  dodgerBlue90: string
  forestGreen: string
  grey: string
  hawkesBlue: string
  lightGrey: string
  nightRider: string
  nightRider50: string
  noble: string
  red: string
  red05: string
  safetyOrange: string
  snow: string
  transparent: string
  whisper: string
  white: string
  whiteSmoke: string
}

export interface NavColorI {
  backgroundHover: string
  color: string
}

export interface FormColorI {
  checkboxCheckedBackground: string
  checkboxCheckedBorder: string
  checkboxColor: string
  checkboxDisabledCheckedBackground: string
  checkboxDisabledCheckedBorder: string
  disabledBackground: string
  disabledBorder: string
  disabledColor: string
  errorBackground: string
  errorBorder: string
  errorColor: string
  focusBackground: string
  focusBorder: string
  placeholderColor: string
}

// This is the color structure for the react-select menus
export interface SelectColorI {
  background: string
  border: string
  cursorColor: string
  disabledBackground?: string
  hover: string
  iconColor: string
  neutral10: string
  optionBorder: string
  primary: string
  primary25: string
  text: string
}

export interface TextColorI {
  copy: string
  copyDark: string
  copyLight: string
  copyWhite: string
  headingDark: string
  headingDefault: string
  headingLight: string
  headingLightest: string
  link: string
  linkActive: string
}

export interface StatusColorI {
  approved: string
  archived: string
  default: string
  draft: string
  pending: string
  rejected: string
}

export interface NotificationColorI {
  default: string
  error: string
  errorBackground: string
  info: string
  success: string
  warning: string
}

export interface ButtonColorI {
  background?: string
  border?: string
  color?: string
}